Add shared speech-text upload preparation
